// src/integrations/supabase/types.ts
export type Database = {
    // Allows to automatically instantiate createClient with right options
    // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
    __InternalSupabase: {
        PostgrestVersion: "13.0.5"
    }
    public: {
        Tables: {
            classes: {
                Row: {
                    category: string
                    created_at: string
                    current_students: number | null
                    description: string
                    id: string
                    image_url: string | null
                    location: string
                    max_students: number | null
                    price: number
                    schedule: string
                    teacher_id: string
                    title: string
                    updated_at: string
                }
                Insert: {
                    category: string
                    created_at?: string
                    current_students?: number | null
                    description: string
                    id?: string
                    image_url?: string | null
                    location: string
                    max_students?: number | null
                    price: number
                    schedule: string
                    teacher_id: string
                    title: string
                    updated_at?: string
                }
                Update: {
                    category?: string
                    created_at?: string
                    current_students?: number | null
                    description?: string
                    id?: string
                    image_url?: string | null
                    location?: string
                    max_students?: number | null
                    price?: number
                    schedule?: string
                    teacher_id?: string
                    title?: string
                    updated_at?: string
                }
                Relationships: [
                    {
                        foreignKeyName: "classes_teacher_id_fkey"
                        columns: ["teacher_id"]
                        isOneToOne: false
                        referencedRelation: "profiles"
                        referencedColumns: ["id"]
                    },
                ]
            }
            profiles: {
                Row: {
                    bio: string | null
                    contact: string | null
                    created_at: string
                    id: string
                    name: string
                    profile_image: string | null
                    role: Database["public"]["Enums"]["user_role"]
                    updated_at: string
                }
                Insert: {
                    bio?: string | null
                    contact?: string | null
                    created_at?: string
                    id: string
                    name: string
                    profile_image?: string | null
                    role?: Database["public"]["Enums"]["user_role"]
                    updated_at?: string
                }
                Update: {
                    bio?: string | null
                    contact?: string | null
                    created_at?: string
                    id?: string
                    name?: string
                    profile_image?: string | null
                    role?: Database["public"]["Enums"]["user_role"]
                    updated_at?: string
                }
                Relationships: []
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            reservation_status: "reserved" | "canceled" | "completed"
            user_role: "teacher" | "student" | "admin"
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
    DefaultSchemaEnumNameOrOptions extends
            | keyof DefaultSchema["Enums"]
        | { schema: keyof DatabaseWithoutInternals },
    EnumName extends DefaultSchemaEnumNameOrOptions extends {
            schema: keyof DatabaseWithoutInternals
        }
        ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
        : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
        schema: keyof DatabaseWithoutInternals
    }
    ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
    : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
        ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
        : never